+++ packages/vtable/src/scenegraph/select/move-select-border.ts
@@ -1,0 +1,12 @@
+import type { Scenegraph } from '../scenegraph';
+
+export function moveSelectingRangeComponentsToSelectedRangeComponents(scene: Scenegraph) {
+  scene.selectingRangeComponents.forEach((rangeComponent, key) => {
+    if (scene.selectedRangeComponents.get(key)) {
+      scene.selectedRangeComponents.get(key).rect.delete();
+    }
+    scene.selectedRangeComponents.set(key, rangeComponent);
+  });
+  scene.selectingRangeComponents = new Map();
+  scene.updateNextFrame();
+}
