+++ packages/vtable/src/scenegraph/utils/get-cell-merge.ts
@@ -1,4 +1,4 @@
-import type { CellRange } from '../../ts-types';
+import type { CellRange, TextColumnDefine } from '../../ts-types';
 import type { BaseTableAPI } from '../../ts-types/base-table';
 
 /**
@@ -9,6 +9,10 @@ import type { BaseTableAPI } from '../../ts-types/base-table';
  * @return {false | CellRange}
  */
 export function getCellMergeInfo(table: BaseTableAPI, col: number, row: number): false | CellRange {
+  // 先判断非表头且非cellMerge配置，返回false
+  if (!table.isHeader(col, row) && (table.getBodyColumnDefine(col, row) as TextColumnDefine).mergeCell !== true) {
+    return false;
+  }
   const range = table.getCellRange(col, row);
   const isMerge = range.start.col !== range.end.col || range.start.row !== range.end.row;
   if (!isMerge) {
